# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/PMWPaymentStatus/verify": {
            "post": {
                "description": "Verifies PMW payment status attestation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Verify PMW Payment Status",
                "parameters": [
                    {
                        "description": "Attestation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attestationtypes.AttestationRequestPMWPaymentStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/attestationtypes.FullAttestationResponsePMWPaymentStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/TeeAvailabilityCheck/verify": {
            "post": {
                "description": "Verifies TEE availability check attestation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Verify TEE Availability Check",
                "parameters": [
                    {
                        "description": "Attestation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attestationtypes.AttestationRequestTeeAvailabilityCheck"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/attestationtypes.FullAttestationResponseTeeAvailabilityCheck"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "attestationtypes.AttestationRequestPMWPaymentStatus": {
            "type": "object",
            "required": [
                "attestationType",
                "requestBody",
                "sourceId"
            ],
            "properties": {
                "attestationType": {
                    "type": "string",
                    "example": "0x4a736f6e41706900000000000000000000000000000000000000000000000000"
                },
                "requestBody": {
                    "$ref": "#/definitions/attestationtypes.IPMWPaymentStatusRequestBody"
                },
                "sourceId": {
                    "type": "string",
                    "example": "0x4a736f6e41706900000000000000000000000000000000000000000000000000"
                }
            }
        },
        "attestationtypes.AttestationRequestTeeAvailabilityCheck": {
            "type": "object",
            "required": [
                "attestationType",
                "requestBody",
                "sourceId"
            ],
            "properties": {
                "attestationType": {
                    "type": "string",
                    "example": "0x4a736f6e41706900000000000000000000000000000000000000000000000000"
                },
                "requestBody": {
                    "$ref": "#/definitions/attestationtypes.ITeeAvailabilityCheckRequestBody"
                },
                "sourceId": {
                    "type": "string",
                    "example": "0x5075626c69635765623200000000000000000000000000000000000000000000"
                }
            }
        },
        "attestationtypes.FullAttestationResponsePMWPaymentStatus": {
            "type": "object",
            "properties": {
                "attestationStatus": {
                    "type": "string"
                },
                "response": {
                    "type": "object",
                    "properties": {
                        "attestationType": {
                            "type": "string"
                        },
                        "requestBody": {
                            "$ref": "#/definitions/attestationtypes.IPMWPaymentStatusRequestBody"
                        },
                        "responseBody": {
                            "$ref": "#/definitions/attestationtypes.IPMWPaymentStatusResponseBody"
                        },
                        "sourceId": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "attestationtypes.FullAttestationResponseTeeAvailabilityCheck": {
            "type": "object",
            "properties": {
                "attestationStatus": {
                    "type": "string"
                },
                "response": {
                    "type": "object",
                    "properties": {
                        "attestationType": {
                            "type": "string"
                        },
                        "requestBody": {
                            "$ref": "#/definitions/attestationtypes.ITeeAvailabilityCheckRequestBody"
                        },
                        "responseBody": {
                            "$ref": "#/definitions/attestationtypes.ITeeAvailabilityCheckResponseBody"
                        },
                        "sourceId": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "attestationtypes.IPMWPaymentStatusRequestBody": {
            "type": "object",
            "properties": {
                "nonce": {
                    "type": "integer"
                },
                "subNonce": {
                    "type": "integer"
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
        "attestationtypes.IPMWPaymentStatusResponseBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "integer"
                },
                "blockTimestamp": {
                    "type": "integer"
                },
                "fee": {
                    "type": "string"
                },
                "paymentReference": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "receivedAmount": {
                    "type": "string"
                },
                "recipientAddress": {
                    "type": "string"
                },
                "revertReason": {
                    "type": "string"
                },
                "senderAddress": {
                    "type": "string"
                },
                "transactionFee": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "transactionStatus": {
                    "type": "integer"
                }
            }
        },
        "attestationtypes.ITeeAvailabilityCheckRequestBody": {
            "type": "object",
            "required": [
                "challenge",
                "teeId",
                "url"
            ],
            "properties": {
                "challenge": {
                    "type": "string",
                    "example": "12345678901234567890"
                },
                "teeId": {
                    "type": "string"
                },
                "url": {
                    "type": "string",
                    "example": "https://supertee.proxy"
                }
            }
        },
        "attestationtypes.ITeeAvailabilityCheckResponseBody": {
            "type": "object",
            "properties": {
                "codeHash": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "initialTeeId": {
                    "type": "string"
                },
                "machineStatus": {
                    "type": "integer"
                },
                "platform": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "rewardEpochId": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "teeGovernanceHash": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "teeTimestamp": {
                    "type": "integer"
                }
            }
        }
    }
}